Added header (AEHeaderView class) with info from current pick set: fantasy team name, owner, owner location, and needs/card count. Used NSAttributedString to highlight needs and card count letters, and center the string (using an attributed string with NSTextField overrides the alignment set in IB).
Fade in and fade out animations for header as cards are animated in and
out.

Added hairline border to scene view so we can see where the edges are
when in full screen mode, or in a window with a non-16:9 aspect.

Fixed problem with team owner name of pick set being pulled from the
wrong array.