Accepting and rejecting player cards from pick sets is fully working. Added symbol for accept or reject and logo to accept or reject remainder when pick set needed count is fulfilled.
Added AEPlayerPickSetStatus enum to AEPlayerPickSet class. Determines
whether a card has been accepted, rejected, or has not yet been
flagged. Needs to be reset when bringing in a new pick set of cards,
since it is stored on the AEPlayer objects, which we are reusing.

Made cards emissive so they aren’t darkened on the outer edges.

Moved adjustment of header labels into the AEFFLHeaderView class
(formerly AEHeader).